+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - json schema transformations are now all multimethod-based.
 - support for `s/eq`
+- support for `nil`
 - BREAKING CHANGE: `type-of` and `return-type-of` are removed in favour of `->json`.
 
 ## 0.7.4 (10.3.2014)