+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - support for `s/eq`
 - support for `nil`
 - BREAKING CHANGE: `type-of` and `return-type-of` are removed in favour of `->json`.
+- BREAKING CHANGE: Route uri is passed as a String (`/api/users/:id`) instead of a Vector (`["/api/users/" :id]`)
 
 ## 0.7.4 (10.3.2014)